// LMS/LMS.UI/huda-center/src/app/shared/models/admin/dashboard-kpi.dto.ts
export interface DashboardKPIDto {
   usersCount: number;
   usersChangePercentage: number;  
   
   employeesCount: number;
   employeesChangePercentage: number;  

   customersCount: number;
   customersChangePercentage: number;  
   newCustomersCount: number;         

   ordersCount: number;               
   ordersChangePercentage: number;    

   booksCount: number;
   booksChangePercentage: number;
   newBooksCount: number;              
}